// src/app/models/user.ts
export class User {
  id: string;
  role: string;
  fullName: string;
  sex: boolean;
  birthday: string;
  active: boolean;
  date: string;
  phone: string;
  email: string;
  password: string;
  token: string;
  passOld: string;
  passConfig: string;
  emailOld: string;
  address: string;
}